backed refactoring and unit tests

+++ backend/WeatherApi.Infrastructure/OpenWeatherMap/Dtos/Coordinates.cs
@@ -5,10 +5,10 @@ namespace WeatherApi.Infrastructure.OpenWeatherMap.Dtos
     public class Coordinates
     {
         [JsonPropertyName("lat")]
-        public float Lat { get; set; }
+        public double Lat { get; set; }
 
         [JsonPropertyName("lon")]
-        public float Lon { get; set; }
+        public double Lon { get; set; }
     }   
 
 }
