+++ backend/WeatherApi.Infrastructure/OpenWeatherMap/Dtos/Wind.cs
@@ -5,7 +5,7 @@ namespace WeatherApi.Infrastructure.OpenWeatherMap.Dtos
     public class Wind
     {
         [JsonPropertyName("speed")]
-        public float Speed { get; set; }
+        public double Speed { get; set; }
 
         [JsonPropertyName("deg")]
         public int Direction { get; set; }
